refactor(ui,types,web): decouple data fetching from presentation across remaining UI domains
Lifts the last deferred-presentation components from packages/web into
@repowise-dev/ui as pure shells, with caller-side wrappers owning data
hooks, mutations, and side effects. Closes the gap from the prior
component-move passes (#119–#125).

Wiki (3): regenerate-button, security-panel, version-history. Plus a
shared MDX components map and a framework-neutral preprocessor so the
RSC wiki-renderer in each consuming app can render without forking the
in-page styling.

Chat (2): chat-interface, conversation-history. Shells take messages /
callbacks via props; the SSE transport, SWR cache, and toast plumbing
stay in the wrapper. Allows alternate transports (e.g. multi-repo
fan-out) to consume the same shell.

Graph (4): graph-flow, graph-doc-panel, graph-community-panel,
path-finder-panel. graph-flow ships as a presentational canvas with
render-slot props for the path-finder + community-panel children;
caller-side wrapper supplies the SWR-driven datasets and renders the
deeper wrappers into those slots. Presentational use-debounce moved to
@repowise-dev/ui/hooks.

Workspace pull-back: package-deps-table promoted to canonical, plus a
slim repo-card-compact variant for callers that don't have RepoStats /
GitSummary per member. WorkspacePackageDepEntry added to canonical
types.

@repowise-dev/types: NodeSearchResult added; GraphPath gains an optional
visual_context; WorkspacePackageDepEntry added.

Verified: packages/{types,ui,web} type-check clean, packages/ui vitest
98/98, packages/web next build clean.
