+++ packages/ui/package.json
@@ -29,6 +29,7 @@
     "./coverage": "./src/coverage/index.ts",
     "./coverage/*": "./src/coverage/*.tsx",
     "./wiki": "./src/wiki/index.ts",
+    "./wiki/wiki-mdx-preprocess": "./src/wiki/wiki-mdx-preprocess.ts",
     "./wiki/*": "./src/wiki/*.tsx",
     "./chat": "./src/chat/index.ts",
     "./chat/*": "./src/chat/*.tsx",
@@ -72,7 +73,9 @@
     "@radix-ui/react-tabs": "^1.1.13",
     "@radix-ui/react-tooltip": "^1.2.8",
     "@repowise-dev/types": "*",
+    "@xyflow/react": "^12.10.2",
     "class-variance-authority": "^0.7.1",
+    "fuse.js": "^7.0.0",
     "clsx": "^2.1.1",
     "d3-force": "^3.0.0",
     "d3-hierarchy": "^3.1.2",
