+++ packages/ui/src/chat/index.ts
@@ -1,7 +1,9 @@
 export * from "./artifact-panel.js";
 export * from "./artifacts.js";
+export * from "./chat-interface.js";
 export * from "./chat-markdown.js";
 export * from "./chat-message.js";
+export * from "./conversation-history.js";
 export * from "./model-selector.js";
 export * from "./source-citation.js";
 export * from "./source-citations.js";
