+++ packages/ui/src/graph/index.ts
@@ -3,6 +3,10 @@ export * from "./graph-ego-sidebar.js";
 export * from "./graph-legend.js";
 export * from "./graph-toolbar.js";
 export * from "./graph-tooltip.js";
+export * from "./graph-doc-panel.js";
+export * from "./graph-community-panel.js";
+export * from "./path-finder-panel.js";
+export * from "./graph-flow.js";
 export * from "./context.js";
 export * from "./elk-layout.js";
 export * from "./use-elk-layout.js";
