+++ packages/ui/src/wiki/index.ts
@@ -2,5 +2,10 @@ export * from "./code-block.js";
 export * from "./confidence-badge.js";
 export * from "./git-history-panel.js";
 export * from "./mermaid-diagram.js";
+export * from "./regenerate-button.js";
+export * from "./security-panel.js";
 export * from "./table-of-contents.js";
+export * from "./version-history.js";
 export * from "./wiki-markdown.js";
+export * from "./wiki-mdx-components.js";
+export * from "./wiki-mdx-preprocess.js";
