+++ packages/ui/src/workspace/index.ts
@@ -2,4 +2,6 @@ export * from "./co-change-table.js";
 export * from "./contract-links-table.js";
 export * from "./contract-type-badge.js";
 export * from "./cross-repo-summary.js";
+export * from "./package-deps-table.js";
 export * from "./repo-card.js";
+export * from "./repo-card-compact.js";
